refactor: 용어 도메인 관련 패키지 변경
