+++ space-d/src/main/java/com/dnd/spaced/core/word/domain/repository/dto/request/WordSearchCondition.java
@@ -1,6 +1,6 @@
 package com.dnd.spaced.core.word.domain.repository.dto.request;
 
-import com.dnd.spaced.core.word.domain.Category;
+import com.dnd.spaced.core.word.domain.enums.Category;
 
 public record WordSearchCondition(String name, Category category, String pronunciation) {
 }
