+++ space-d/src/test/java/com/dnd/spaced/config/docs/snippet/DocsController.java
@@ -11,8 +11,8 @@ import com.dnd.spaced.core.account.domain.enums.ProfileImageName;
 import com.dnd.spaced.core.quiz.domain.enums.QuizCategory;
 import com.dnd.spaced.core.report.domain.enums.ReportReason;
 import com.dnd.spaced.core.report.domain.enums.ReportStatus;
-import com.dnd.spaced.core.word.domain.Category;
-import com.dnd.spaced.core.word.domain.PronunciationType;
+import com.dnd.spaced.core.word.domain.enums.Category;
+import com.dnd.spaced.core.word.domain.enums.PronunciationType;
 import com.dnd.spaced.global.exception.code.AccountErrorCode;
 import com.dnd.spaced.global.exception.code.AuthErrorCode;
 import com.dnd.spaced.global.exception.code.BookmarkErrorCode;
